+++ Garten_Simulation/Views/ScreenTimeSuggestionsView.swift
@@ -1,0 +1,7 @@
+import SwiftUI
+
+struct ScreenTimeSuggestionsView: View {
+    var body: some View {
+        Text("Suggestions")
+    }
+}